# api/openapi.yaml
openapi: 3.0.3
info:
  title: Routewerk API
  version: 0.1.0
  description: |
    REST API for Routewerk's competition tracking and supporting endpoints.
    This is the source of truth for both the Go server (`internal/api/`,
    via `oapi-codegen`) and the SvelteKit client (`web/spa/src/lib/api/`,
    via `openapi-typescript`). Run `make api-gen` to regenerate after
    spec changes; CI fails if the committed generated files drift.

    Phase 1e populates magic-link auth and comp CRUD as a thin end-to-end
    slice that exercises the full pipeline (spec → Go types → TS types →
    SPA consumption). Phase 1f adds the remaining comp endpoints (events,
    problems, registrations, the unified action endpoint, verification,
    and the leaderboard + SSE stream).
servers:
  - url: https://routewerk.fly.dev
    description: production
  - url: https://routewerk-dev.fly.dev
    description: staging
  - url: http://localhost:8080
    description: local dev
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /competitions/{id}/registrations:
    parameters:
      - $ref: '#/components/parameters/CompetitionId'
    get:
      tags: [competitions]
      operationId: listRegistrations
      summary: List registrations for a competition
      description: |
        Any authenticated user can list. The leaderboard-visibility
        filter (wave 4) governs how much detail unaffiliated users see.
      responses:
        '200':
          description: Registrations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompetitionRegistration'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [competitions]
      operationId: createRegistration
      summary: Register self (or another climber, if staff) for a competition
      description: |
        With no `user_id` field, registers the calling user. Staff
        (`gym_manager+` at the comp's location) may pass a `user_id`
        to register someone else.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCreate'
      responses:
        '201':
          description: Registration created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionRegistration'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Climber already registered for this competition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registrations/{id}/attempts:
    parameters:
      - $ref: '#/components/parameters/RegistrationId'
    get:
      tags: [competitions]
      operationId: listRegistrationAttempts
      summary: Per-problem state for a registration
      description: |
        Returns the climber's current per-problem attempt state, in the
        same shape the action endpoint emits. Used by the SPA scorecard
        to hydrate on first load (and after a refresh).

        Caller must be the registration's user OR `gym_manager+` at the
        comp's location.
      responses:
        '200':
          description: Per-problem attempt state.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttemptState'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /registrations/{id}:
    parameters:
      - $ref: '#/components/parameters/RegistrationId'
    delete:
      tags: [competitions]
      operationId: withdrawRegistration
      summary: Withdraw a registration (mark withdrawn)
      description: |
        Caller must be the registration's user OR `gym_manager+` at the
        comp's location. Soft delete — the row stays for audit, the bib
        number is freed for reuse.
      responses:
        '204':
          description: Withdrawn.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Action endpoint (the league-night centerpiece) ─────────────
...
components:
  parameters:
...
    CompetitionId:
      name: id
      in: path
      required: true
      description: Competition UUID.
      schema:
        type: string
        format: uuid
...
    RegistrationId:
      name: id
      in: path
      required: true
      description: Competition registration UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request body or invalid arguments.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message. Matches the existing routewerk envelope.
      example:
        error: "competition not found"
...
    CompetitionRegistration:
      type: object
      required: [id, competition_id, category_id, user_id, display_name, created_at]
      properties:
        id: { type: string, format: uuid }
        competition_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        display_name:
          type: string
          description: Snapshot at registration time.
        bib_number: { type: integer, nullable: true }
        waiver_signed_at: { type: string, format: date-time, nullable: true }
        paid_at: { type: string, format: date-time, nullable: true }
        withdrawn_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    RegistrationCreate:
      type: object
      required: [category_id]
      properties:
        category_id: { type: string, format: uuid }
        user_id:
          type: string
          format: uuid
          description: |
            Optional. Defaults to the calling user. Setting a different
            user_id requires `gym_manager+` role at the comp's location.
        display_name:
          type: string
          description: Optional. Defaults to the user's display_name.
        bib_number:
          type: integer
          nullable: true
          description: Staff-assigned bib. Climbers self-registering should leave null.

    # ── Action endpoint ─────────────────────────────────────────
...
    AttemptState:
      type: object
      required: [problem_id, attempts, zone_reached, top_reached]
      properties:
        problem_id: { type: string, format: uuid }
        attempts: { type: integer }
        zone_attempts: { type: integer, nullable: true }
        zone_reached: { type: boolean }
        top_reached: { type: boolean }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        For mobile / API clients. Obtain via `POST /api/v1/auth/login`
        (returns `access_token` + `refresh_token`). Header:
        `Authorization: Bearer <jwt>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: _rw_session
      description: |
        For browser / SPA clients. Set automatically after a successful
        password login (`POST /login`) or magic-link verify
        (`GET /verify-magic`). HttpOnly + SameSite=Lax + Secure (in prod).